Rename 'Primitives' in the Domain and 'Common' in the Infrastructure to 'Shared', plus move 'Behaviors' in the Application to the new 'Shared' namespace
