+++ account-management/src/Application/Shared/Behaviors/ValidationPipelineBehavior.cs
@@ -1,7 +1,7 @@
 using FluentValidation;
 using MediatR;
 
-namespace PlatformPlatform.AccountManagement.Application.Behaviours;
+namespace PlatformPlatform.AccountManagement.Application.Shared.Behaviors;
 
 public sealed class ValidationPipelineBehavior<TRequest, TResponse> : IPipelineBehavior<TRequest, TResponse>
     where TRequest : IRequest<TResponse>
