+++ account-management/src/Domain/Tenants/ITenantRepository.cs
@@ -1,4 +1,4 @@
-using PlatformPlatform.AccountManagement.Domain.Primitives;
+using PlatformPlatform.AccountManagement.Domain.Shared;
 
 namespace PlatformPlatform.AccountManagement.Domain.Tenants;
